Factor out test server into conftest.py
To allow reuse across test cases
